+++ src/game/tests.rs
@@ -611,6 +611,7 @@ mod held_tapped;
 mod howling_mine;
 mod identity_and_mana;
 mod instill_energy;
+mod into_the_wilds;
 mod isd_dgm_deck_cards;
 mod isd_token_cards;
 mod jade_statue;